Try to continue after smtp errors
Current behavior aborts when getting an error from the smtp server when sending feedback. This is problematic when it's a user specific error (ie we have the wrong address), because it crashes after sending some (but not all) feedbacks, and worse yet, the order it sends them doesn't match the excel file, so you have to go through and remove students line by line in order to resend. Much easier to just fail for specific users.